Рассчет обрезанного среднего в long

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -25,9 +25,9 @@ public class SalesManager {
         return min;
     }
 
-    public int getAvarageSales() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long getAvarageSales() {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
